# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flora-Chloris Flowershop POS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface UpdateProfileRequest {
  name?: string;
  phone?: string | null;
  email?: string | null;
  currentPassword?: string | null;
  newPassword?: string | null;
}

export interface RegisterRequest {
  username: string;
  password: string;
  name: string;
  phone?: string | null;
  email?: string | null;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  cashier: "cashier",
  customer: "customer",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  name: string;
  phone?: string | null;
  email?: string | null;
}

export interface LoginResponse {
  user: User;
  message: string;
}

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  emoji?: string | null;
}

export interface CreateCategoryRequest {
  name: string;
  description?: string | null;
  emoji?: string | null;
}

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  purchaseCost?: number | null;
  categoryId: number;
  categoryName?: string | null;
  stockQuantity: number;
  isAvailable: boolean;
  imageUrl?: string | null;
  unit?: string | null;
}

export interface CreateProductRequest {
  name: string;
  description?: string | null;
  price: number;
  purchaseCost?: number | null;
  categoryId: number;
  stockQuantity: number;
  imageUrl?: string | null;
  unit?: string | null;
  isAvailable?: boolean;
}

export interface UpdateProductRequest {
  name?: string;
  description?: string | null;
  price?: number;
  purchaseCost?: number | null;
  categoryId?: number;
  stockQuantity?: number;
  isAvailable?: boolean;
  imageUrl?: string | null;
  unit?: string | null;
}

export interface InventoryItem {
  productId: number;
  productName: string;
  categoryName?: string | null;
  stockQuantity: number;
  isAvailable: boolean;
  price: number;
  unit?: string | null;
  imageUrl?: string | null;
}

export interface UpdateInventoryRequest {
  stockQuantity: number;
  isAvailable?: boolean;
}

export interface OrderItem {
  productId: number;
  productName?: string | null;
  quantity: number;
  unitPrice: number;
  subtotal: number;
}

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  walk_in: "walk_in",
  online: "online",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  type: OrderType;
  status: OrderStatus;
  customerName?: string | null;
  customerPhone?: string | null;
  items: OrderItem[];
  subtotal: number;
  total: number;
  paymentMethod?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateOrderRequestType =
  (typeof CreateOrderRequestType)[keyof typeof CreateOrderRequestType];

export const CreateOrderRequestType = {
  walk_in: "walk_in",
  online: "online",
} as const;

export type CreateOrderRequestItemsItem = {
  productId: number;
  quantity: number;
};

export interface CreateOrderRequest {
  type: CreateOrderRequestType;
  customerName?: string | null;
  customerPhone?: string | null;
  items: CreateOrderRequestItemsItem[];
  paymentMethod?: string | null;
  notes?: string | null;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
  paymentMethod?: string | null;
}

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ReservationPaymentMethod =
  | (typeof ReservationPaymentMethod)[keyof typeof ReservationPaymentMethod]
  | null;

export const ReservationPaymentMethod = {
  cod: "cod",
  gcash: "gcash",
} as const;

export type ReservationFulfillmentType =
  (typeof ReservationFulfillmentType)[keyof typeof ReservationFulfillmentType];

export const ReservationFulfillmentType = {
  pickup: "pickup",
  delivery: "delivery",
} as const;

export interface Reservation {
  id: number;
  reservationNumber: string;
  userId?: number | null;
  customerName: string;
  customerPhone: string;
  customerEmail?: string | null;
  items: OrderItem[];
  pickupDate: string;
  status: ReservationStatus;
  total: number;
  notes?: string | null;
  paymentMethod?: ReservationPaymentMethod;
  fulfillmentType: ReservationFulfillmentType;
  createdAt: string;
}

export type CreateReservationRequestItemsItem = {
  productId: number;
  quantity: number;
};

export type CreateReservationRequestPaymentMethod =
  | (typeof CreateReservationRequestPaymentMethod)[keyof typeof CreateReservationRequestPaymentMethod]
  | null;

export const CreateReservationRequestPaymentMethod = {
  cod: "cod",
  gcash: "gcash",
} as const;

export type CreateReservationRequestFulfillmentType =
  (typeof CreateReservationRequestFulfillmentType)[keyof typeof CreateReservationRequestFulfillmentType];

export const CreateReservationRequestFulfillmentType = {
  pickup: "pickup",
  delivery: "delivery",
} as const;

export interface CreateReservationRequest {
  customerName: string;
  customerPhone: string;
  customerEmail?: string | null;
  items: CreateReservationRequestItemsItem[];
  pickupDate: string;
  notes?: string | null;
  paymentMethod?: CreateReservationRequestPaymentMethod;
  fulfillmentType?: CreateReservationRequestFulfillmentType;
}

export type UpdateReservationStatusRequestStatus =
  (typeof UpdateReservationStatusRequestStatus)[keyof typeof UpdateReservationStatusRequestStatus];

export const UpdateReservationStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateReservationStatusRequest {
  status: UpdateReservationStatusRequestStatus;
}

export interface DashboardStats {
  totalSalesToday: number;
  totalOrdersToday: number;
  pendingReservations: number;
  lowStockItems: number;
  totalSalesThisMonth: number;
  totalOrdersThisMonth: number;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  available?: boolean;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  type?: ListOrdersType;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListOrdersType =
  (typeof ListOrdersType)[keyof typeof ListOrdersType];

export const ListOrdersType = {
  walk_in: "walk_in",
  online: "online",
} as const;

export type ListReservationsParams = {
  status?: ListReservationsStatus;
};

export type ListReservationsStatus =
  (typeof ListReservationsStatus)[keyof typeof ListReservationsStatus];

export const ListReservationsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;