# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Flora-Chloris Flowershop POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated or wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, cancelled]
        - name: type
          in: query
          schema:
            type: string
            enum: [walk_in, online]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (POS or online)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        currentPassword:
          type: string
          nullable: true
        newPassword:
          type: string
          nullable: true
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      required:
        - username
        - password
        - name
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, cashier, customer]
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      required:
        - id
        - username
        - role
        - name
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required:
        - productId
        - quantity
        - unitPrice
        - subtotal
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        type:
          type: string
          enum: [walk_in, online]
        status:
          type: string
          enum: [pending, completed, cancelled]
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - type
        - status
        - items
        - subtotal
        - total
        - createdAt
    CreateOrderRequest:
      type: object
      properties:
        type:
          type: string
          enum: [walk_in, online]
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - productId
              - quantity
        paymentMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - type
        - items
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, completed, cancelled]
        paymentMethod:
          type: string
          nullable: true
      required:
        - status
...